requirements/common.txt
Django==3.0.6
google-api-python-client==1.8.2
youtube_dl==2020.5.8
PyVimeo==1.0.11
djangorestframework==3.11.0
arrow==0.15.6
factory-boy==2.12.0
psycopg2-binary==2.*,>=2.8.4
opencv-python==4.2.0.34
django-storages==1.9.1
